fix: :bug: update all prvious path to capture the new codebase defination
